requirements.txt
django==3.1
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
dj-database-url==0.5.0
psycopg2==2.8.6
whitenoise===5.2.0
djangorestframework_simplejwt==0.4.1

